+++ include/Ark/VM/VM.hpp
@@ -35,6 +35,51 @@ namespace Ark
         void loadFunction(const std::string& name, internal::Value::ProcType function);
         void run();
 
+        internal::Value& operator[](const std::string& name);
+
+        template <typename... Args>
+        internal::Value&& call(const std::string& name, Args&&... args)
+        {
+            using namespace Ark::internal;
+
+            // find id of function
+            auto it = std::find(m_symbols.begin(), m_symbols.end(), name);
+            if (it == m_symbols.end())
+            {
+                if constexpr (debug)
+                    throwVMError("Couldn't find symbol with name " + name);
+            }
+
+            // find function object and push it if it's a pageaddr/closure
+            uint16_t id = static_cast<uint16_t>(std::distance(m_symbols.begin(), it));
+            auto var = findNearestVariable(id);
+            if (var != nullptr)
+            {
+                if (var->valueType() != ValueType::PageAddr && var->valueType() != ValueType::Closure)
+                    throwVMError("Symbol " + name + " isn't a function");
+                
+                push(*var);
+                m_last_sym_loaded = id;
+            }
+            else
+                throwVMError("Couldn't load symbol with name " + name);
+
+            // convert and push arguments in reverse order
+            std::vector<Value> fnargs { args... };
+            for (auto it2=fnargs.rbegin(); it2 != fnargs.rend(); ++it2)
+                push(*it2);
+
+            std::size_t frames_count = m_frames.size();
+            // call it
+            call(static_cast<int16_t>(sizeof...(Args)));
+
+            // run until the function returns
+            safeRun(/* untilFrameCount */ frames_count);
+
+            // get result
+            return pop();
+        }
+
     private:
         bool m_persist;
         bytecode_t m_bytecode;
@@ -44,6 +89,7 @@ namespace Ark
         bool m_running;
         std::string m_filename;
         uint16_t m_last_sym_loaded;
+        std::size_t m_until_frame_count;
 
         // related to the bytecode
         std::vector<std::string> m_symbols;
@@ -61,6 +107,7 @@ namespace Ark
         // bytecode releated
 
         void configure();
+        void safeRun(std::size_t untilFrameCount=0);
 
         inline uint16_t readNumber()
         {
@@ -127,6 +174,9 @@ namespace Ark
             }
 
             m_frames.back().resetScopeCountToDelete();
+
+            if (m_frames.size() == m_until_frame_count)
+                m_running = false;
         }
 
         inline void createNewScope()
@@ -160,7 +210,7 @@ namespace Ark
         inline void popJumpIfFalse();
         inline void jump();
         inline void ret();
-        inline void call();
+        inline void call(int16_t argc_=-1);
         inline void capture();
         inline void builtin();
         inline void mut();